* Refactor deleting savegames
At least for the "Load game" part this should work like before.
What bothers me at the moment is that deleting savegames seems to not
work as intended: For me, unlink() will not complain about deleting the
files, but an empty file of the same name is still retained somehow.